attempt to fix Int cannot be dereferenced error

+++ src/com/amazon/kinesis/streaming/agent/tailing/AbstractParser.java
@@ -118,7 +118,7 @@ public abstract class AbstractParser<R extends IRecord> implements IParser<R> {
                         totalUndhandledErrors.incrementAndGet();
                         return false;
                     }
-                    sleep((1000.0 * pow(1.7, retryCount.doubleValue())).longValue());
+                    sleep((1000.0 * pow(1.7, retryCount)).longValue());
                 }
             }
         }
